refactor: nest NamespaceDependencyFilterResult in Filtered.Types and address analyzer hints
- Move NamespaceDependencyFilterResult to a nested type of Filtered.Types,
  consistent with the StringEqualityResult pattern
- SignatureDependencyCollector: return HashSet from Build, move Unwrap/
  SafeAttributes/SafeAttributeArguments inside the collector (only used there),
  and return List from SafeAttributeArguments
